+++ packages/plugins/FileService/src/SNSAdaptor/modals/modals.ts
@@ -1,6 +1,6 @@
 import { SingletonModal } from '@masknet/shared-base'
 import type { ConfirmModalOpenProps } from './ConfirmModal/index.js'
-import type { RenameModalOpenProps } from './RenameDialog/index.js'
+import type { RenameModalCloseProps, RenameModalOpenProps } from './RenameModal/index.js'
 
 export const ConfirmModal = new SingletonModal<ConfirmModalOpenProps>()
-export const RenameModal = new SingletonModal<RenameModalOpenProps>()
+export const RenameModal = new SingletonModal<RenameModalOpenProps, RenameModalCloseProps>()
